Remove support for non-threadsafe interfaces.
Fixes #449.

With the v0.9.0 release of UniFFI, non-threadsafe interfaces are explicitly
deprecated. This commit removes support for non-threadsafe interfaces entirely,
making threadsafety the default and only supported mode. In practice, for
any `interface` declared in the UDL:

* Its implementation must be `Sync` and `Send`.
* Its methods cannot accept a mutable `&self` parameter.

The `[Threadsafe]` attribute may still be specified explicitly in the UDL
file, but will result in a deprecation warning and is otherwise ignored.

We don't have a good way to write tests for the following, but I've manually
verified that:

* A non-`Send+Sync` interface implementation produces a compile-time error.
* Use of the `[Threadsafe]` attribute results in a deprecation warning.